saliency heatmaps debugging

+++ icairdweakly/create_saliency_heatmaps.py
@@ -275,14 +275,15 @@ if __name__ == '__main__':
             for n in range(num_classes):
                 print(F.interpolate(hipe_maps[n],(pdim, pdim)).shape)
                 full_hipe_maps[n][x: x1, y:y1] = F.interpolate(hipe_maps[n],(pdim, pdim))[0][0]
-                full_hipe_seg[x:x1, y:y1] = F.interpolate(hipe_seg.unsqueeze(0).unsqueeze(0), (pdim, pdim))[0][0]
+                full_hipe_seg[x:x1, y:y1] = F.interpolate(hipe_seg.float().unsqueeze(0).unsqueeze(0), (pdim,
+                                                                                                       pdim))[0][0]
 
         wandb.log({
             'Full HiPe'             : [wandb.Image(full_hipe_maps[h], caption=label_list[h]) for h in range(
                     num_classes)],
             'Full HiPe Segmentation': wandb.Image(full_img, masks={
                 "predictions": {
-                    "mask_data": full_hipe_seg.numpy(), "class_labels": class_labels
+                    "mask_data": full_hipe_seg.int().numpy(), "class_labels": class_labels
                     }
                 })
             })
